code/requirements.txt
numpy==1.21.2
pytorch==1.10.1
torchvision==0.11.2
cudatoolkit
tqdm==4.62.3
optuna==2.10.0
pandas==1.4.0
matplotlib==3.5.1
opencv==4.5.5
